# Performance improvement `r_lgl_which()`