Originally incorrectly created in https://github.com/r-lib/vctrs/pull/1594.

``` r
set.seed(1)

r_lgl_which <- function(x, na_propagate) {
  stopifnot(rlang::is_logical(x), rlang::is_bool(na_propagate))
  .Call(rlang:::ffi_test_lgl_which, x, na_propagate)
}

n <- 100e6
all_true <- vctrs::vec_rep(TRUE, n)
all_false <- vctrs::vec_rep(FALSE, n)
true_90 <- runif(n) < 0.9
true_50 <- runif(n) < 0.5
true_10 <- runif(n) < 0.1
true_1 <- runif(n) < 0.01

bench::mark(
  all_true = r_lgl_which(all_true, FALSE),
  all_false = r_lgl_which(all_false, FALSE),
  true_90 = r_lgl_which(true_90, FALSE),
  true_50 = r_lgl_which(true_50, FALSE),
  true_10 = r_lgl_which(true_10, FALSE),
  true_1 = r_lgl_which(true_1, FALSE),
  check = FALSE,
  min_iterations = 10
)
# Current dev
#> # A tibble: 6 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 all_true    142.8ms  148.5ms      5.08  381.47MB    3.38 
#> 2 all_false    91.1ms   96.1ms     10.5    39.26KB    0    
#> 3 true_90     222.9ms  237.5ms      3.97  343.32MB    0.992
#> 4 true_50     548.3ms  629.6ms      1.65  190.73MB    0.413
#> 5 true_10     210.5ms  224.6ms      4.45   38.14MB    0    
#> 6 true_1      106.5ms  109.1ms      9.14    3.82MB    0

# This PR
#> # A tibble: 6 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 all_true    113.6ms  118.2ms      6.47  381.47MB    2.77 
#> 2 all_false    53.3ms   54.7ms     17.8     1.95MB    0    
#> 3 true_90     203.6ms  207.5ms      4.82  381.47MB    2.06 
#> 4 true_50     518.9ms  538.7ms      1.82  199.13MB    0.454
#> 5 true_10     170.1ms  181.8ms      5.52   54.93MB    0.613
#> 6 true_1       67.7ms   71.4ms     14.1     9.13MB    0
```

<sup>Created on 2022-09-21 with [reprex v2.0.2](https://reprex.tidyverse.org)</sup>

Currently, this PR is only a proof of concept. It uses some magic numbers I just came up with without any testing at all. Maybe it would also make sense to not have a fixed jump size to avoid some very bad estimates, e.g. a vector of length `200 * n` where every second element is true would break the estimate.

I haven't yet added tests. Ideally, a test would force resizing of the vector but to construct such an input depends on the details of the implementation.