@mgirlich 2022-09-21
Comparing the performance of the "all false" case to the PR in vctrs it seems that we might pay quite a bit for the `has_names` check in every iteration.